Update for better Mac support
On Macs you have an error that will show up. I belive I've gotten it fixed but for better experince I've got it setup to catch the error and log it to the console (if present). In addition Its setup to exit on close if its not windows for now.